backend/pyproject.toml
[project]
name = "chatbot-platform-backend"
version = "0.1.0"
description = "Multi-user chatbot platform backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=43.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "openai>=1.54.0",
    "anthropic>=0.39.0",
    "python-multipart>=0.0.12",
    "bcrypt>=4.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langgraph>=0.2.0",
    "langchain-community>=0.3.0",
    "python-jose[cryptography]>=3.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
